Stored user and related files inside folder
