--- Admin.h
@@ -1,12 +1,0 @@
-#include "User.h"
-
-class Admin : public User
-{
-public:
-	Admin();
-	~Admin();
-
-private:
-
-};
-
