--- Client.cpp
@@ -1,9 +1,0 @@
-#include "Client.h"
-
-Client::Client()
-{
-}
-
-Client::~Client()
-{
-}
