--- Client.h
@@ -1,10 +1,0 @@
-#include "User.h"
-#include "Account.h"
-
-class Client : public User
-{
-public:
-	Client();
-	~Client();
-private:
-};
