+++ Data Structures/Admin.h
@@ -1,0 +1,12 @@
+#include "User.h"
+
+class Admin : public User
+{
+public:
+	Admin();
+	~Admin();
+
+private:
+
+};
+
