+++ Data Structures/Client.cpp
@@ -1,0 +1,9 @@
+#include "Client.h"
+
+Client::Client()
+{
+}
+
+Client::~Client()
+{
+}
