+++ Data Structures/Client.h
@@ -1,0 +1,10 @@
+#include "User.h"
+#include "Account.h"
+
+class Client : public User
+{
+public:
+	Client();
+	~Client();
+private:
+};
