+++ Data Structures/Official.cpp
@@ -1,0 +1,11 @@
+#include "Official.h"
+
+string Official::getState()
+{
+	return officialState;
+}
+
+void Official::setState(string newState)
+{
+	officialState = newState;
+}
