--- Official.cpp
@@ -1,11 +1,0 @@
-#include "Official.h"
-
-string Official::getState()
-{
-	return officialState;
-}
-
-void Official::setState(string newState)
-{
-	officialState = newState;
-}
