--- Official.h
@@ -1,11 +1,0 @@
-#include "User.h"
-
-class Official : public User
-{
-public:
-	string getState();
-	void setState(string newState);
-
-private:
-	string officialState;
-};
